+++ lib/Console/Command/CrawlCommand.php
@@ -41,6 +41,7 @@ class CrawlCommand extends Command
     private const OPT_EXCLUDE_URL = 'exclude-url';
     private const OPT_HEADER = 'header';
     private const OPT_RATE = 'rate';
+    private const OPT_INCLUDE_LINK = 'include-link';
 
     /**
      * @var DispatcherBuilderFactory
@@ -96,6 +97,7 @@ class CrawlCommand extends Command
         $this->addOption(self::OPT_EXCLUDE_URL, null, InputOption::VALUE_REQUIRED|InputOption::VALUE_IS_ARRAY, 'Exclude PCRE URL pattern', []);
         $this->addOption(self::OPT_HEADER, null, InputOption::VALUE_REQUIRED|InputOption::VALUE_IS_ARRAY, 'Custom header, e.g. "X-Teapot: Me"', []);
         $this->addOption(self::OPT_RATE, null, InputOption::VALUE_REQUIRED, 'Set max request rate (as requests per second)', []);
+        $this->addOption(self::OPT_INCLUDE_LINK, null, InputOption::VALUE_REQUIRED|InputOption::VALUE_IS_ARRAY, 'Add an additional URL to the set of URLs under the base URL', []);
     }
 
     protected function execute(InputInterface $input, OutputInterface $output)
@@ -162,6 +164,7 @@ class CrawlCommand extends Command
         $excludeUrls = $this->castToArray($input->getOption(self::OPT_EXCLUDE_URL));
         $headers = $this->castToArray($input->getOption(self::OPT_HEADER));
         $rate = $input->getOption(self::OPT_RATE);
+        $includeLinks = $this->castToArray($input->getOption(self::OPT_INCLUDE_LINK));
         
         $builder = $this->factory->createForUrls($urls);
         $builder->maxConcurrency($maxConcurrency);
@@ -196,6 +199,10 @@ class CrawlCommand extends Command
 
         if (!empty($excludeUrls)) {
             $builder->excludeUrlPatterns($excludeUrls);
+        }
+
+        if (!empty($includeLinks)) {
+            $builder->includeLinks($includeLinks);
         }
 
         return $builder->build();